backend/requirements.txt
fastapi==0.109.0

uvicorn==0.27.0

google-generativeai==0.3.0

python-dotenv==1.0.0

pydantic==2.9.0

supabase==2.0.1

bcrypt==4.1.2

websockets>=12.0,<13.0